Degraded mode write IO: - Moved SNS repair process query out of VERSION_MISMATCH if condition. - Moved corresponding SNS repair process query code to be unconditional   in io_bottom_half(). - Used U32 instead of U64 for repair done member in IO reply fop. - Changed SNS repair process query API to return a tri-state instead   of a boolean.

+++ ioservice/io_fops.h
@@ -374,10 +374,11 @@ struct m0_fop_cob_rw_reply {
 	/**
 	 * A field indicating whether repair has finished or not for given
 	 * global fid.
-	 * rwr_repair_done == 1 indicates file has been repaired.
+	 * rwr_repair_done == -1 indicates SNS repair has not started at all.
 	 * rwr_repair_done == 0 indicates file is still be to be repaired.
+	 * rwr_repair_done == 1 indicates file has been repaired.
 	 */
-	uint64_t             rwr_repair_done;
+	uint32_t             rwr_repair_done;
 } M0_XCA_RECORD;
 
 /**
